requirements.txt
langchain
google-auth
google-auth-oauthlib
google-api-python-client
streamlit
openai
aiohttp
mcp
typing-extensions